Extract menu items to separated components
